Change Lesson8. (I deleted previous brunch, because it has junk artifact). 1. Change variable name from Triangle.class 2. Delete empty space Triangle.class 3. Rename packages. 4. Delete variable count previous element.
